# Conflicts:
#	src/SiteCohorts.cs

Add cohort.LastLAI to cloning of initialSites
Change LAIPerSpecies to float, use LastLAI as source of cohortLAI
Limit foliage, LAI allocation if FolTentative > 0.01 to avoid calculation of 0 LAI

+++ src/ISiteCohorts.cs
@@ -47,7 +47,7 @@ namespace Landis.Library.PnETCohorts
 
         Landis.Library.Parameters.Species.AuxParm<int> NSCPerSpecies { get; }
 
-        Landis.Library.Parameters.Species.AuxParm<int> LAIPerSpecies { get; }
+        Landis.Library.Parameters.Species.AuxParm<float> LAIPerSpecies { get; }
 
         Landis.Library.Parameters.Species.AuxParm<int> WoodySenescencePerSpecies { get; }
